fix: rename relay to recoil
