+++ src/containers/WeekViewer/index.tsx
@@ -4,7 +4,7 @@ import { useRecoilValue } from 'recoil';
 
 import { DayViewer, YearAndMonth } from '../../components';
 
-import { habitDay } from '../../relay/atoms';
+import { habitDay } from '../../recoil/atoms';
 
 import { getWeek, isSameDate } from '../../utils';
 import { Week } from '../../types';
